Added controller tests
