+++ config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
     namespace :api do
       namespace :v1 do
         resources :categories, only: %i[index create]
-        resources :articles, only: %i[index create]
+        resources :articles, except: %i[edit new], param: :slug
       end
     end
   end
