+++ test/controllers/api/v1/categories_controller_test.rb
